\docType{data}
\name{clinical_data}
\alias{clinical_data}
\title{Clinical data}
\format{Matrix}
\description{
  This matrix contains clinical data (age, gender, type II diabetes status and BMI)
  of the 506 FGENTCARD patients from whom the metabolic profiles ("metabolic_data" dataset)
  were obtained.
}
\usage{data(clinical_data)}
\value{Matrix}
