\docType{data}
\name{metabolic_data}
\alias{metabolic_data}
\title{NMR plasma metabolic profiles}
\format{Matrix}
\description{
  This matrix contains the 1H NMR profiles (1.60 - 0.80 ppm) of 506 plasma samples
  from the FGENTCARD cohort and 10 identical quality control (QC) samples. The QC samples
  were prepared from a representative pool of the experimental samples, and were injected
  regularly throughout the run to ensure analytical reproducibility.
}
\usage{data(metabolic_data)}
\value{Matrix}
