\docType{data}
\name{target_metabolites}
\alias{target_metabolites}
\title{NMR profiles of 9 target metabolites}
\format{Matrix}
\description{
  This matrix contains the 1H NMR profiles of 9 target metabolites (lactate, 3-hydroxy-
  butyrate, leucine, valine, isoleucine, N-acetylated compounds, acetate, alanine
  and 1,5-anhydroglucitol) across the experimental samples and the quality control
  (QC) samples.
}
\usage{data(target_metabolites)}
\value{Matrix}
